First pass at generating a row and the page to show it in
JS file with functions to generate a row, check if all elements are the
same and a function that returns a string value based on what select
elements are worth. Also a simulate button that simulates a set number
of spins.
A html page that has a button to spin, check and return the results to
the screen.